"extras" for Text
This allows users to add arbitrary attributes to a Text without losing
the semblance of security that instantiating a Text gives via
json_decode (otherwise, a user could slip in arbitrary double underscore
keys into the __dict__, which could be problematic).

It may be nice to add search functionality over information stored in
"extras", but that is an idea for the future.