Bug fix: no double counting; need at least two match features
Suppose the following two units:
 * u1:  A B C D A
 * u2:  A A Q Q X
In this case, u1[0] will match with u2[0] and u2[1], and u1[4] will
match with u2[0] and u2[1].  We do not want to double count u2[0] and
u2[1] during scoring, even though they both match with both u1[0] and
u1[4].  That is the no double counting fix.

More importantly, the above will not be counted as a match because they
only match on feature A; there needs to be at least two different
features matched in order to count the pair of units as a match.

These two fixes bring us closer to the list of matches v3 produces on
the Lucan-Vergil benchmark.