Bug fix: insert arbitrary number of tokens
Because regular insert does a lookup check to the database with all
tokens as argument to finding duplicates, attempting to insert too many
tokens at once will lead to a DocumentTooLarge error.  While batching
the insertions would have been one way to handle this problem, I decided
to make use of the insert method that does not make the duplication
check.