Faster _get_trivial_distance
While a single call to _get_trivial_distance is close to the microsecond
level, this function gets called rather often.  Refactoring made this
function twice as fast, which ultimately shaved off a few tenths of a
second in the Lucan-Vergil benchmark, but every little bit adds up to
bigger savings.