Greek2latin (#74)
* Data loading module for greek to latin

* Include greek to latin data with distributed package

* Script for initial database setup

The entries of the synonym dictionaries will need to be pre-registered
with the database in order for them to work with the current ingestion
and searching algorithms.

This initial setup script also allows for indexing the database.

* Refactor feature index retrieval for stopwords

* Refactor in anticipation of Greek to Latin

* Refactor tests

Preparing to add Greek to Latin test

* Add Greek to Latin tests

* Refactor tests for database calculations

* (Tests failing) initial implementation of Greek to Latin

Scores are wrong, which means frequency calculation is wrong somewhere

* Fixed truth values

Forgot to give v3 scores enough precision. Tests now pass.

* Refactor calculations testing

In preparation for testing Greek to Latin frequency calculations

* More refactoring

* Test for greek to latin frequencies on greek text

* Corrected Greek to Latin frequency calculation

* Stub out greek to latin with corpus frequencies

Because of the way Greek to Latin is currently implemented, it is
difficult to get the same frequency information as v3. Perhaps this will
be something to work on later.

* Increased precision for v3 scores in tests

* Fixed tests so that matches are properly retrieved

* Further tests fixing

* Greek to Latin corpus frequencies correct?

The test passed

* Mold GreekToLatinSearch to fit infrastructure

* Make tests run